Abstracted ChronoSyncedDataStructure
